% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextToPlot.R
\name{addTextToPlot}
\alias{addTextToPlot}
\title{Add Text to a Plot Using Relaive Positoning}
\usage{
addTextToPlot(x, y, text, ...)
}
\arguments{
\item{x}{Numeric value of horizontal positon from 0 (left) to 1 (right)}

\item{y}{Numeric value of vertical positon from 0 (bottom) to 1 (top)}

\item{text}{Character string to be written on the plot.}

\item{...}{Additional arguments passed to text().}
}
\value{
None
}
\description{
Easy little way to add a text to a plot.  The plot area is rescaled to range from 0 to 1 in both directions.
Text placement is on this scale with larger values of y and x moving the placement up and to the right respectively.\cr
}
\examples{
addTextToPlot(0.5, 0.5, text="X", )  #X marks the spot, in this case the center)
addTextToPlot(0.95, 0.05, text="y = mx + b", adj=1) #right justified
}
\author{
Gordon W. Holtgrieve
}
