% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcuateK0.R
\name{CalculateK0}
\alias{CalculateK0}
\title{Calculates Henery's constant (mol/L/atm)}
\usage{
CalculateK0(temperature, gas, salinity = 0.25, pressure = 1)
}
\arguments{
\item{temperature}{Water temperature in degrees Celsius.}

\item{salinity}{Salinity in practical salinity units (PSU). Defaults to 0.}

\item{pressure}{Surface pressure in atmospheres. Defaults to 1.}
}
\value{
Numeric vector of Henry's constant for a specified gas in units of mol/L/atm.
}
\description{
Calculates Henry's constant for multiple gases at a given temperature, salinity and
pressure. Currently supported gases include CO2, CH4, O2, Ar, N2, and N2O.  Units are converted from
kg to volume (L) using water density given by the rho() function in seacarb.
}
\note{
Units of mol/L/atm and umol/L/uatm are equivalent.
}
\author{
Gordon W. Holtgrieve (gholt@uw.edu)
}
