% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.isotope.standard.name.R
\name{check.isotope.standard.name}
\alias{check.isotope.standard.name}
\title{Checks given names of an isotope standard against a list of possible names.}
\usage{
check.isotope.standard.name(x)
}
\arguments{
\item{x}{Proposed name to be checked against list of possible names.}
}
\value{
Character vector contaiing the 'offical' name of the standard.
}
\description{
Convenience function for checking a given name of an isotope standard against a list of possible names,
returning the 'offical' name for use going forward.
}
\author{
Gordon W. Holtgrieve
}
