% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.to.R
\name{delta.to.R}
\alias{delta.to.R}
\title{Converts standard delta notation to isotopic ratio}
\usage{
delta.to.R(x, std = NULL)
}
\arguments{
\item{x}{Numeric.}

\item{std}{Character vector of length one indicating the accepted international standard the data are relative to.
           Possible options are:
\describe{
  \item{VPDB}{13C/12C = 0.011180 (note: carbon only)}
  \item{air-N}{15N/14N = 0.003676}
  \item{air-O}{18O/16O = 0.00205292}
  \item{VSMOW-O}{18O/16O = 0.00200520}
  \item{VSMOW-H}{D/H = 0.00015576}
  \item{VCDT}{34S/32S = 0.0441626}
}}
}
\value{
Object same as 'x' with data as an isotopic ratio.
}
\description{
This function is used to convert isotope data in standard delta notation into isotopic ratio (R).
If argument 'std' is provided, then the raw ratio (R = heavy isotope / light isotope) is returned.
Do not use this option if ratios are given as light isotope / heavy isotope (not typical).  If 'std'
is NULL (default) then the 'grand ratio' (R-sample/R-standard) is returned.
}
\examples{
delta.to.R(10, std = "air")  # Returns R_sample
delta.to.R(1.0)  # Returns R_sample/R_standard
}
\author{
Gordon W. Holtgrieve
}
