% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.adjust.R
\name{EA.adjust}
\alias{EA.adjust}
\title{Corrects raw d13C and 15N to vs. VPBD and air}
\usage{
EA.adjust(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\value{
List containing four dataframes:
  \describe{
    \item{standard.CN}{Dataframe of raw data IsoDat data for the standards and QTY standards; updated to with new column for final corrected data.}
    \item{sample.CN}{Dataframe of raw data IsoDat data for the unknown samples form the sequence; updated to with new column for final corrected data.}
    \item{calibration.coefficients}{Dataframe of slope and intercept values used during data standardization.}
    \item{measured.standard.means}{Dataframe of average d13C and d15N of standards vs. tank and vs. international standards.}
   }

standard C & N data ($standard.CN),
        sample C & N data ($sample.CN),
        and calibration model coefficients ($calibration.coefficients).
}
\description{
This function takes the raw d13C and d15N from IsoDat software and corrects them to VPDB and air, respectively,
             based on the working standards from the run. Two-point calibration curves are generated based on measured and known
             values of the working standards; this calibration is then applied to both the samples and standards.
             The function also calculates the percent C and N of samples and standards in a similar fashion using known
             mass C and N in the glutamic acid standard (GA1) and peak area (masses 44 and 28).
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
