% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.blank.correct.R
\name{EA.blank.correct}
\alias{EA.blank.correct}
\title{Blank corrects EA sample data. (Not Implemented)}
\usage{
EA.blank.correct(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\description{
(Currently Not Implemented) Analyzes blanks for the run and reports result to the user, then asking whether or not to
              blank correct standard and sample data. The correction is based on empty tins and a simple mixing model.
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
