% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.check.R
\name{EA.check}
\alias{EA.check}
\title{Calculates precision and accuracy metrics of finalized data from NACHO}
\usage{
EA.check(results)
}
\value{
List with one tibble:
   \describe{
     \item{error.analysis.results}{Table containing accuracy and precision of derived data.}
   }
}
\description{
Calculates precision and accuracy of d13C, d15N, mass percent C, and mass percent N.
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
