% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.drift.correct.R
\name{EA.drift.correct}
\alias{EA.drift.correct}
\title{Drift corrects raw data from the EA based on standards}
\usage{
EA.drift.correct(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\value{
List with two tables and a Boolean flag:
  \describe{
    \item{standard.CN}{Dataframe of raw data IsoDat data for the standards and QTY standards; updated to with new column for drift corrected data, if required.}
    \item{sample.CN}{Dataframe of raw data IsoDat data for the unknown samples form the sequence; updated to with new column for drift corrected data, if required.}
    \item{drift.correct.flag}{Boolean flag to indicate if a drift correction was performed.}
    }
}
\description{
This function uses repeated analysis of known standard to check for an, if desired, correct raw data for instrumental drift.
             The function first test for sufficient data to make the correction. If data are insufficient, the raw data ar copied
             unchanged to new column, .csv written, and outcome printed to the console. If the data are sufficient, a drift analysis
             is done and plots generated. The user is asked if they want to drift correct the C data, N data, or both. The function
             returns a character vector that indicates the results of the drift correct as  N, C, both, or none.
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
