% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.load.data.R
\name{EA.load.data}
\alias{EA.load.data}
\title{XXX}
\usage{
EA.load.data(results)
}
\arguments{
\item{results}{List containing XXX}
}
\value{
List with X objects...
}
\description{
This function XXX
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
