% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.NACHO.R
\name{EA.NACHO}
\alias{EA.NACHO}
\title{Finalizes raw EA data from NACHO}
\usage{
EA.NACHO(data.file.dir = getwd(), combine.runs = F, area.cutoff = F)
}
\arguments{
\item{data.file.dir}{Character vector of length 1 containing file path to raw .csv files.}

\item{combine.runs}{Boolean flag identifying if the standard across all runs in the folder should be combined into a single standard curve or if each run should be analyzed independently.
FALSE means each run is corrected independently. Default is to combine standards (==T).}

\item{area.cutoff}{Defines the peak area (Vs) cutoff below which an individual analysis (injection) is dropped.  FALSE means no cutoff applied.  Numeric defines the cutoff value.}

\item{data.file}{Character vector of full file names for raw data files from the instrument. Must be a .csv file!}
}
\value{
Dataframe of data for each unknown sample that includes following finalized values (i.e., what you want...):
  \describe{
    \item{d13C.vs.VPDB}{13C:12C of bulk carbon in the sample in delta notation with units of per mil relatve to Vienna Pee Dee Belemite.}
    \item{d15N.vs.air}{15N:14N of bulk nitrogen in the sample in delta notation with units of per mil relatve to atmospheric air.}
    \item{pctC}{Percent carbon in the sample on a mass basis.}
    \item{pctN}{Percent nitrogen in the sample on a mass basis.}
   }
}
\description{
This function is used to decompose and finalize EA data from solid samples analyzed on the Delta V irMS systems (NACHO).
The function writes a .csv with the thinned raw data and a .pdf with all relevant information about the analysis.  The finalized
sample data are written as a .csv file and a dataframe of the data is also returned.

The input data file has strict formatting and data requirements.  (This all needs updating) Each sample or standard should have four rows of data: two for  N and
2 C.  Deviations for the is will result in the function failing.  The follow columns must be present in the input data file for the function to work.
  \describe{
    \item{Row}{Row number for the IsoDat sequence file.}
    \item{Analysis}{Unique number assigned by IsoDat for each analysis.}
    \item{Identifier 1}{Unique identifier for each sample or standard in the file. If two runs are combined, be sure this field remains unique. Character.}
    \item{Comment}{Identifies whether the particular analysis is a 'STD', 'QTY', 'SAMPLE', 'BLANK', 'ZERO', or 'DUMMY'. These are the only vaild options for this field. Character.}
    \item{Amount}{Mass of SAMPLE, STANDARD, QTY, or DUMMY in miligrams (mg). BLANK and ZERO can be 0.0 or blank. Numeric.}
    \item{Is Ref _}{Flag that identifies wheter this line line is for a reference gas peak (==1) or not (==0). Numeric.}
    \item{Area 28}{Peak area for mass 28 in Vs. Numeric.}
    \item{d 15N/14N}{Ratio of 15N to 14N in delta units relative to the refernce gas (i.e., N2 tank). Numeric.}
    \item{Area 44}{Peak area for mass 44 in Vs. Numeric.}
    \item{d 13C/12C}{Ratio of 13C to 12C in delta units relative to the refernce gas (i.e., CO2 tank). Numeric.}
    }

The remaining columns should be unchanged from what is created by IsoDat.  There should be a total of 49 columns of data in the raw data file.
Column names will be modified when imported to R.
}
\author{
Gordon W. Holtgrieve
}
