% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.plot.standards.R
\name{EA.plot.standards}
\alias{EA.plot.standards}
\title{Plots peak area vs. mass and d13C/d15N vs. peak area of working standards.}
\usage{
EA.plot.standards(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\value{
List containing lm objects from four linear models:
             mass N vs. m/z 28
             mass C vs. m/z 44
             d15N vs. m/z 28
             d13C vs. m/z 44
}
\description{
This function...
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
