% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.report.R
\name{EA.report}
\alias{EA.report}
\title{Write final data file and generate report}
\usage{
EA.report(results)
}
\arguments{
\item{data.file}{Character vector of full file names for raw data files from the instrument. Must be a .csv file!}
}
\description{
This internal function ...
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
