% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.standards.R
\name{EA.standards}
\alias{EA.standards}
\title{Determines relevant standards for the run.}
\usage{
EA.standards(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\value{
List with three (3) objects:
  \describe{
    \item{standard.CN}{Dataframe of raw data IsoDat data for the standards and QTY standards; updated to include known mass C, mass N, d13C, and d15N.}
    \item{known.standard.values}{Dataframe of known values of the working standards used.}
    \item{zero.CN}{Character vector of standards used in this data set.}
    }
}
\description{
This internal function determines the standards used in a run and reports that back to the user + saves for later use.
It also calculates the expected mass C and N of the standards based on the based on measured weight (amount) and known percent C or N.
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
