% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exetainers.NACHO.R
\name{exetainers.NACHO}
\alias{exetainers.NACHO}
\title{Finalizes extetainer gas ratio data from NACHO}
\usage{
exetainers.NACHO(data.file, analysis.date, area.cutoff = F, injections.standards = c(F, T, T, T), injections.samples = c(F, T, T, T), lab.air.T = 20, system.pressure.atm = 1.49701, salinity = 0)
}
\arguments{
\item{data.file}{Full file name for raw data from the instrument. Must be a .csv file! Character.}

\item{analysis.date}{Date of analysis on NACHO as a character.}

\item{area.cutoff}{Defines the peak area (Vs) cutoff below which an individual analysis (injection) is droped.  FALSE means no cutoff applied.  Numeric defines the cutoff value.}

\item{injections.standards}{Boolean vector length 4 idenitfying which injections of the standards to drop (F) or keep (T) for use in the calculations.}

\item{injections.samples}{Boolean vector length 4 idenitfying which injections of the samples to drop (F) or keep (T) for use in the calculations.}

\item{lab.air.T}{Temperature of the lab at time of analysis in degrees Celcius.  Defaults to 20.  Numeric.}

\item{system.pressure.atm}{Pressure of the inlet system in atmospheres.  Defaults to 1.49701 (22 psi).  Numeric.}

\item{salinity}{Salninty of the sample in ppt.  Defaults to 0.  Numeric.}
}
\value{
Dataframe of :
  \describe{
    \item{}{}
   }
}
\description{
This function is used to decompose and finalize gas ratio data from exetainers analyzed on the Delta V irMS systems (NACHO).  The first step
is thin and organige the raw data.  Plots are generated (as .pdfs) to look for mass effects on the results ansd the user is asked whether to correct for
sample mass differences among injections and samples.  After corrections, measured values are calibrated against the working standards included
in the run.  The function writes a .csv with the thinned raw data and a .txt with all relevant information about the analysis.  The finalized
sample data are writen as a .csv file and adataframe the data is also returned.
}
\author{
Gordon W. Holtgrieve
}
