% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.K0.R
\name{get.K0.CO2}
\alias{get.K0.CO2}
\title{Calculates Henry's constant for CO2 (mol/L/atm)}
\usage{
get.K0.CO2(temperature = 20, salinity = 0, pressure.atm = 1)
}
\arguments{
\item{temperature}{Water temperature in degrees Celsius.}

\item{salinity}{Salinity in practical salinity units (PSU). Defaults to 0.}

\item{pressure.atm}{Surface pressure in atmospheres. Defaults to 1.}
}
\value{
Numeric vector of Henry's constant for a specified gas in units of mol/L/atm.
}
\description{
Calculates Henry's constant for CO2 at a given temperature, salinity and pressure.
}
\note{
Units of mol/L/atm and umol/L/uatm are equivalent.
}
\author{
Gordon W. Holtgrieve
}
