% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg.to.umol.R
\name{mg.to.umol}
\alias{mg.to.umol}
\title{Converts mass or concentration data from mass to molar units}
\usage{
mg.to.umol(x, mol)
}
\arguments{
\item{x}{Mass (or concentration) of the molecule, in milligrams, to be converted}

\item{mol}{Character vector of length one indicating the molecule in question.}
}
\value{
Object same as 'x' but in units of micromoles
}
\description{
This function is used to convert mass or concentration data from units of milligrams to micromoles.
}
\examples{
mg.to.umol(52.8114, mol="CO2")
}
\author{
Gordon W. Holtgrieve
}
