% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.to.delta.R
\name{R.to.delta}
\alias{R.to.delta}
\title{Converts isotopic ratio data to standard delta notation}
\usage{
R.to.delta(x, isotope.system=NULL, reference.std = NULL)
}
\arguments{
\item{x}{Numeric.}

\item{isotope.system}{Character vector of length one indicating the (single) isotope system of interest.
Currently supported isotope systems are "C", "N", "S", "H", "O18", "O17".}

\item{reference.std}{Character vector of length one indicating the reference standard the data are relative to.}
}
\value{
Object same as 'x' with data in delta notation rounded to two decimal places.
}
\description{
This function is used to convert isotopic ratio data (R) into standard delta notation.
If arguments 'isotope.system' and 'reference.std' are provided, then x should be the raw ratio (R = heavy isotope / light isotope).
Do not use this option if ratios are given as light isotope / heavy isotope (not typical). If 'isotope.system' or 'reference.std' is
NULL (default) then x should be the 'grand ratio' (R-sample/R-standard).
}
\author{
Gordon W. Holtgrieve
}
