% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umol.to.mg.R
\name{umol.to.mg}
\alias{umol.to.mg}
\title{Converts moles or molar concentration data from molar units to mass units}
\usage{
umol.to.mg(x, mol)
}
\arguments{
\item{x}{Mass (or concentration) of the molecule, in micromoles, to be converted}

\item{mol}{Character vector of length one indicating the molecule in question.}
}
\value{
Object same as 'x' but in units of milligrams
}
\description{
This function is used to moles or molar concentration data from units of micromoles to miligrams.
}
\examples{
umol.to.mg(1200, mol = "CO2")
}
\author{
Gordon W. Holtgrieve
}
