% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterIso.Picarro.R
\name{waterIso.Picarro}
\alias{waterIso.Picarro}
\title{Finalizes water isotope data from a Picarro CRDS}
\usage{
waterIso.Picarro(data.file, sample.file, working.STDs = c("BW","KD","SW"))
}
\arguments{
\item{data.file}{Full file name for raw data from the instrument. Must be a .csv file! File name must start with the instrument ID number follow by "_". Character.}

\item{sample.file}{Full file name for the sample information that corresponds to the raw data.  Also, must be a .csv file! The file must include the
                 following headings (and only the following headings).  Best to use the template!
\describe{
  \item{trayNum}{The tray positon on the autosampler.}
  \item{vialNum}{The vail number of the sample.}
  \item{sampleType}{Aceptable identifiers include: Conditioner, Standard, and Sample.}
  \item{SampleDesc}{Unique identifier of the sample.}
  }}

\item{working.STDs}{Character vector of length 3 listing names of working stadards included in the run. The first two will be used for correction and calibration,
the third as the QA/QC check.  Defaults to IsoLab standards "BW", "KD", "SW".}
}
\value{
Dataframe of finalized water isotope data including:
  \describe{
    \item{trayNum}{The tray positon on the autosampler.}
    \item{vialNum}{The vail number of the sample.}
    \item{SampleDesc}{Unique identifier of the sample.}
    \item{H2O.ppmv.mean}{Average water concentration across saved injections in ppm by volume.}
    \item{H2O.ppmv.SD}{Standard deviation of water concentration across saved injections in ppm by volume.}
    \item{dD.mean.VSMOW}{Average 2H:1H of water across saved injections in standard delta notation relative to VSMOW.}
    \item{dD.SD}{Standard deviation of 2H:1H across saved injections in standard delta notation relative to VSMOW.}
    \item{d18O.mean.VSMOW}{Average 18O:16O of water across saved injections in standard delta notation relative to VSMOW.}
    \item{d18O.SD}{Standard deviation of 18O:16O across saved injections in standard delta notation relative to VSMOW.}
   }
}
\description{
This function is used to decompose and finalize raw water isotope data from a Picarro CRDS instrument.  The first step
is to correct for linear drift of the instrument.  The second step calibrate the measured values against the working standards included
in the run and calculate QA/QC metrics.  The function writes a .txt with all relevant information included.  A dataframe of
the finalized sample data is also returned.
}
\author{
Gordon W. Holtgrieve
}
