% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.to.R
\name{delta.to.R}
\alias{delta.to.R}
\title{Converts standard delta notation to isotopic ratio}
\usage{
delta.to.R(x, isotope.system=NULL, reference.std = NULL)
}
\arguments{
\item{x}{Numeric.}

\item{isotope.system}{Character vector of length one indicating the (single) isotope system of interest.
Currently supported isotope systems are "C", "N", "S", "H", "O18", "O17".}

\item{reference.std}{Character vector of length one indicating the reference standard the data are relative to.}
}
\value{
Object same as 'x' with data as an isotopic ratio.
}
\description{
This function is used to convert isotope data in standard delta notation into isotopic ratio (R).
If arguments 'isotope.system' and 'reference.std' are provided, then the raw ratio (R = heavy isotope / light isotope) is returned.
Do not use this option if ratios are given as light isotope / heavy isotope (not typical).
If 'isotope.system' or 'reference.std' is NULL (default) then the 'grand ratio' (R-sample/R-standard) is returned.
}
\author{
Gordon W. Holtgrieve
}
