% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.blank.correct.R
\name{EA.blank.correct}
\alias{EA.blank.correct}
\title{Blank corrects EA sample data. (Not Implemented)}
\usage{
EA.blank.correct(results)
}
\value{
List with two tables: standard C & N data ($standard.CN),
        sample C & N data ($sample.CN), and a flag indicating if a
        blank correction was performed ($blank.correct.flag).
}
\description{
(Currently Not Implemented) Analyzes blanks for the run and reports result to the user, then asking whether or not to
              blank correct standard and sample data. The correction is based on empty tins and a simple mixing model.
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
