% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.check.peak.areas.R
\name{EA.check.peak.areas}
\alias{EA.check.peak.areas}
\title{Check peak areas}
\usage{
EA.check.peak.areas(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\value{
Boolean dataframe
}
\description{
This internal function check whether the peak areas of samples (masses 28 and 44) are 30%, 50% or 300% of
the target mass. Target masses are 200 ug C (mass 44) and 40 ug N (mass 38).
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
