% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.check.R
\name{EA.check}
\alias{EA.check}
\title{Calculates precision and accuracy metrics of finalized data from NACHO}
\usage{
EA.check(results)
}
\arguments{
\item{results}{List containing results from previous functions.}
}
\value{
List with two objects: known.standard.values, error.analysis.results
}
\description{
This function calculates the r
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
