% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EA.load.data.R
\name{EA.load.data}
\alias{EA.load.data}
\title{Loads raw EA data for further analysis.}
\usage{
EA.load.data(results)
}
\arguments{
\item{results}{List containing XXX}
}
\value{
List with X objects...
}
\description{
This internal function loads raw EA data for analysis and makes it available for further analysis. If more than one
data file name (and path) is supplied, those files will e combined into a single dataframe and returned. The function also asks
the user to identify nteractively the directory to place the processed data file and report.
}
\author{
Gordon W. Holtgrieve
}
\keyword{internal}
