% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exetainers.NACHO.R
\name{exetainers.NACHO}
\alias{exetainers.NACHO}
\title{Finalizes extetainer gas ratio data from NACHO}
\usage{
exetainers.NACHO(data.file, analysis.date, area.cutoff = F, injections.standards = c(F, T, T, T), injections.samples = c(F, T, T, T), lab.air.T = 20, system.pressure.atm = 1.49701, salinity = 0)
}
\arguments{
\item{data.file}{Full file name for raw data from the instrument. Must be a .csv file! Character.}

\item{area.cutoff}{Defines the peak area (Vs) cutoff below which an individual analysis (injection) is droped.  FALSE means no cutoff applied.  Numeric defines the cutoff value.}

\item{injections.standards}{Boolean vector length 4 idenitfying which injections of the standards to drop (F) or keep (T) for use in the calculations.}

\item{injections.samples}{Boolean vector length 4 idenitfying which injections of the samples to drop (F) or keep (T) for use in the calculations.}

\item{lab.air.T}{Temperature of the mass spec lab at time of analysis in degrees Celcius.  Defaults to 20.  Numeric.}

\item{system.pressure.atm}{Pressure of the inlet system in atmospheres.  Defaults to 1.49701 (22 psi).  Numeric.}

\item{salinity}{Salninty of the sample in ppt.  Defaults to 0.  Numeric.}
}
\value{
Dataframe of finalized data for each exeatiner including:
  \describe{
    \item{d180_02.vs.air}{18O:16O of diatomic oxygen (O2) in the headspace in standard delta notation relative to atmospheric air.}
    \item{d180_02.vs.VSMOW}{18O:16O of diatomic oxygen (O2) in the headspace in standard delta notation relative to VSMOW.}
    \item{d170_02.vs.air}{17O:16O of diatomic oxygen (O2) in the headspace in standard delta notation relative to atmospheric air.
          Use with caution as this method has not been validated.}
    \item{d15N_N2.vs.air}{15N:14N of diatomic nitrogen (N2) in the headspace in standard delta notation relative to atmospheric air.
          Use with caution as this method has not been validated.}
    \item{O2.Ar}{Oxygen to argon ratio (mass 32:40) of headspace gas.}
    \item{d13C.CO2}{13C:12C of carbon dioxide (CO2) in the headspace (after acidificatioon) in standard delta notation relative to VPDB.}
    \item{d13C.TotalDIC}{13C:12C of dissolved inroganic carbon in standard delta notation relative to VPDB calculated from the headspace
          CO2 and using Henry's constant to estimate the 13C:12C of CO2 that raims dissoved in water.  Note that it may be necessary to
          further correct these data based on co-analyzed DIC standards.}
   }
}
\description{
This function is used to decompose and finalize gas ratio data from exetainers analyzed on the Delta V irMS systems (NACHO).  The first step
is thin and organige the raw data.  Plots are generated (as .pdfs) to look for mass effects on the results ansd the user is asked whether to correct for
sample mass differences among injections and samples.  After corrections, measured values are calibrated against the working standards included
in the run.  The function writes a .csv with the thinned raw data and a .txt with all relevant information about the analysis.  The finalized
sample data are writen as a .csv file and a dataframe of the data is also returned.

The input data file has strick formatting and data requirments.  Each sample or standard should have seven rows of data: 3 for massess 32, 34
and 40; 2 for masses 28 and 29, and 2 for masses 44, 45, 46.  Deviations for the is will result in the function failing.  In additon the first four columns
must be the following four identification columns in this order:
  \describe{
    \item{analysisDate}{Data of the run(s) on NACHO. Character.}
    \item{Identifier.1}{Unique identifier for each exetainer in the file. If two runs are combined, be sure this field remains unique. Character.}
    \item{Identifier.2}{Mililiters of air added to air standards OR volume of headspace in the sample or water standard vial. Headspace volume is the difference
          of the inital, full vial weigth and weight after generating the headspace. Numeric.}
    \item{Comment}{Identifies whether the extetainer is a 'airSTD', 'waterSTD', 'Sample' or 'Conditioner'. These are the only vaild options for this field. Character.}
  }
The remaining columns should be unchaged from what is created by IsoDat.  There should be a total of 45 columsn of data in the raw data file.
}
\author{
Gordon W. Holtgrieve
}
