% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.isotope.standard.R
\name{get.isotope.standard}
\alias{get.isotope.standard}
\title{Helper function to access information in the isotope.standards dataframe}
\usage{
get.isotope.standard(std, isotope.system)
}
\arguments{
\item{std}{Caharcter vector with name(s) of the isotope standard of interest.}

\item{isotope.system}{Character vector of length one indicating the (single) isotope system of interest. Currently supported isotope systems are "C", "N", "S", "H", "O18", "O17".}
}
\description{
This function is used to pull out isotope standard data from the isotope.standards dataframe
easily and efficently.  The isotope.standards dataframe contains only raw isotope ratios,
so this function also will convert to delta value relative to a specified standard.
}
\author{
Gordon W. Holtgrieve
}
