% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.molar.mass.R
\name{get.molar.mass}
\alias{get.molar.mass}
\title{Calculate average molar mass of molecules.}
\usage{
get.molar.mass(mol)
}
\arguments{
\item{mol}{Character vector of simple chemical formulas.}
}
\value{
Numeric vector of average molar masses same length as 'mol'.
}
\description{
Convenience function for obtaining the average molar mass of chemical compounds as found in nature.
}
\author{
Gordon W. Holtgrieve
}
